tools.py
## Importing libraries and files
import os
from dotenv import load_dotenv
load_dotenv()

# Simple functions that CrewAI can use
def read_financial_document(file_path: str = 'data/sample.pdf') -> str:
    """Tool to read and extract text from a PDF financial document"""
    try:
        # For now, let's use a simple text reader
        if not os.path.exists(file_path):
            return f"File not found: {file_path}. Please upload a financial document."
        
        # Simple file reading - we'll improve this later
        try:
            with open(file_path, 'rb') as f:
                # Basic text extraction placeholder
                return f"PDF file loaded: {file_path}. Content analysis would be performed here."
        except Exception as e:
            return f"Could not read PDF: {str(e)}. Please ensure it's a valid PDF file."
            
    except Exception as e:
        return f"Error accessing file: {str(e)}"

def search_tool(query: str) -> str:
    """Basic search tool placeholder"""
    return f"Search functionality: Would search for '{query}' in financial databases and return relevant market data, news, and analysis."
