requirements.txt
click
crewai
crewai-tools
fastapi
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
Jinja2
jsonschema
langchain-core
langchain-community
langchain-openai
langsmith
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
pandas
pillow
pip
protobuf
pydantic
pydantic_core
python-dotenv
python-multipart
pypdf
uvicorn
